+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+- \[engine, types\] Make timeout parameter optional for conditional transfer and withdraw
+
 ## 0.2.0
 
 - \[router\] Rebalance from Matic -> ETH Mainnet